// src/app/data-access/models/device.model.ts
export enum DeviceStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  ALL = "ALL",
}

//Device Network
export enum DeviceNetwork {
  lor = "lor",
  N01 = "N01",
  R234 = "R234",
  sig = "sig",
  uncl = "uncl",
  ALL = "ALL",
}
// ...
export enum WirelessStatus {
  WiFi = "wifi",
  RSSI = "RSSI",
  SIGFOX = "sigfox",
  LORA = "lora",
  NB = "nb-iot",
  SNR = "SNR",
}

export interface IiconData {
  icon: string[];
  color: string;
}